CORE-13796: Move the long lived objects into an internal TracingState object
